Move OpInfo to core org.scijava.ops package
It's the central data structure for metadata about an op.
