+++ imagej/imagej-ops2/src/main/java/net/imagej/ops2/help/AbstractHelp.java
@@ -33,9 +33,9 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
+import org.scijava.ops.OpInfo;
 import org.scijava.ops.OpUtils;
 import org.scijava.ops.matcher.OpCandidate;
-import org.scijava.ops.matcher.OpInfo;
 
 /**
  * Base class for help operations.
