+++ scijava/scijava-ops/src/main/java/org/scijava/ops/OpEnvironment.java
@@ -29,7 +29,6 @@
 
 package org.scijava.ops;
 
-import org.scijava.ops.matcher.OpInfo;
 import org.scijava.ops.matcher.OpRef;
 
 /**
