+++ scijava/scijava-ops/src/main/java/org/scijava/ops/OpUtils.java
@@ -39,7 +39,6 @@ import java.util.stream.Collectors;
 import org.scijava.ops.matcher.MatchingResult;
 import org.scijava.ops.matcher.OpCandidate;
 import org.scijava.ops.matcher.OpCandidate.StatusCode;
-import org.scijava.ops.matcher.OpInfo;
 import org.scijava.ops.matcher.OpMatcher;
 import org.scijava.ops.matcher.OpRef;
 import org.scijava.param.ParameterMember;
