+++ scijava/scijava-ops/src/main/java/org/scijava/ops/function/OpWrappers.java
@@ -9,8 +9,8 @@ import java.lang.reflect.Type;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 
-import org.scijava.ops.matcher.OpInfo;
 import org.scijava.types.GenericTyped;
+import org.scijava.ops.OpInfo;
 import org.scijava.ops.util.OpWrapper;
 import org.scijava.param.Mutable;
 import org.scijava.plugin.Plugin;
