+++ scijava/scijava-ops/src/main/java/org/scijava/ops/matcher/OpCandidate.java
@@ -36,6 +36,7 @@ import java.util.Map;
 
 import org.scijava.log.Logger;
 import org.scijava.ops.OpEnvironment;
+import org.scijava.ops.OpInfo;
 import org.scijava.ops.OpUtils;
 import org.scijava.param.ValidityProblem;
 import org.scijava.struct.Member;
