+++ scijava/scijava-ops/src/main/java/org/scijava/ops/matcher/OpFieldInfo.java
@@ -38,6 +38,7 @@ import java.util.List;
 
 import org.scijava.Priority;
 import org.scijava.ops.OpField;
+import org.scijava.ops.OpInfo;
 import org.scijava.ops.OpUtils;
 import org.scijava.param.ParameterStructs;
 import org.scijava.param.ValidityException;
